+++ maiko/src/internal.rs
@@ -1,5 +1,7 @@
 mod actor_controller;
 mod broker;
+mod command;
+mod command_sender;
 mod step_handler;
 mod step_pause;
 mod subscriber;
@@ -7,6 +9,8 @@ mod subscription;
 
 pub(crate) use actor_controller::ActorController;
 pub(crate) use broker::Broker;
+pub(crate) use command::Command;
+pub(crate) use command_sender::CommandSender;
 pub(crate) use step_handler::StepHandler;
 pub(crate) use step_pause::StepPause;
 pub(crate) use subscriber::Subscriber;